# opt: count number of execPlan columns correctly