Fixes #32723.

Prior to this commit, we assumed that the number of columns output by
an execPlan was equal to the size of its outputCol map. This was untrue
in the case where we output a single column multiple times, which could
happen when we reuse projections in a query like `SELECT '', ''`.

I found two places where this was clearly causing a bug:
1. when we initialize a scalarBuildCtx we sized the ivarHelper to be the
   size of the outputCols map. This led to it being undersized and causing
   an out-of-bounds in some cases.
2. When we built a projectSet we used this count to offset our
   projections, leading to incorrect results.

We now take the highest value (column index) in the map to be the number
of output columns.

Here's before/after numbers for ExecBuild:
```
name                                 old time/op    new time/op    delta
Phases/kv-read/ExecBuild-8             12.0µs ± 2%    11.7µs ± 2%  -2.19%  (p=0.000 n=10+10)
Phases/kv-read-no-prep/ExecBuild-8     41.2µs ± 2%    42.2µs ± 4%  +2.49%  (p=0.003 n=9+10)
Phases/kv-read-const/ExecBuild-8        662ns ± 2%     660ns ± 2%    ~     (p=0.753 n=10+10)
Phases/tpcc-new-order/ExecBuild-8      41.8µs ± 3%    41.8µs ± 2%    ~     (p=0.968 n=10+9)
Phases/tpcc-delivery/ExecBuild-8       27.8µs ± 2%    28.4µs ± 2%  +2.38%  (p=0.001 n=10+10)
Phases/tpcc-stock-level/ExecBuild-8     141µs ± 0%     141µs ± 2%    ~     (p=0.965 n=8+10)
```

Release note (bug fix): Fixed a panic in the cost-based optimizer during
the planning of some queries.